Remove duplicate group files replaced by univeral
